+++ src/Notifications/LeadAssignedNotification.php
@@ -2,12 +2,16 @@
 
 namespace Goldnead\Leadhub\Notifications;
 
+use Goldnead\Leadhub\Contracts\BrandAddressed;
 use Goldnead\Leadhub\Models\Contact;
+use Goldnead\Leadhub\Notifications\Concerns\SendsAsBrand;
 use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
 
-class LeadAssignedNotification extends Notification
+class LeadAssignedNotification extends Notification implements BrandAddressed
 {
+    use SendsAsBrand;
+
     public function __construct(public Contact $contact) {}
 
     public function via(object $notifiable): array
@@ -19,7 +23,7 @@ class LeadAssignedNotification extends Notification
     {
         $name = $this->contact->displayName();
 
-        return (new MailMessage)
+        $message = (new MailMessage)
             ->subject(__('leadhub::notifications.assigned.subject', ['name' => $name]))
             ->greeting(__('leadhub::notifications.assigned.greeting'))
             ->line(__('leadhub::notifications.assigned.line', ['name' => $name]))
@@ -27,5 +31,7 @@ class LeadAssignedNotification extends Notification
                 __('leadhub::notifications.view_contact'),
                 cp_route('leadhub.contacts.show', $this->contact->uuid),
             );
+
+        return $this->asBrand($message);
     }
 }